Changed a few defaults and added a new option.
Changed the default seperator to '>' for easier use with "cut".

Removed the ability to change the string wrapper in order to leverage
the strconv.Quote functionality which will escape all of the special
characters in a string in order to ensure everything stays on one line.

Now hiding the keys by default. Added an option to display them.